+++ local-development.php
@@ -14,7 +14,7 @@
  * Author:            Andy Fragen
  * Author URI:        http://thefragens.com/
  * Description:       Places development notice for plugins or themes that are in local development. Prevents updating of selected plugins and themes. Automatically adds plugins and themes under version control. Allows for using a local file server and bypassing the WordPress 5.2 WSOD Shutdown Handler.
- * Version:           2.6.2
+ * Version:           2.7.0
  * Domain Path:       /languages
  * Text Domain:       local-development
  * License:           GNU General Public License v2
@@ -50,4 +50,4 @@ if ( version_compare( phpversion(), '5.6', '<=' ) ) {
 
 // Load Autoloader.
 require_once __DIR__ . '/vendor/autoload.php';
-( new Bootstrap() )->run();
+( new Bootstrap() )->run( __FILE__ );