+++ CHANGES.md
@@ -1,5 +1,11 @@
 #### [unreleased]
 
+#### 2.7.0 / 2020-099-15
+* set custom admin bar to display environment type
+* remove `WP_ENVIRONMENT_TYPE` on deactivation
+* remove a line of jQuery for removing `update` row class as unnecessary and now [conflicting with GHU](https://github.com/afragen/github-updater/pull/898)
+* disable the auto-update link from WP 5.5
+
 #### 2.6.2 / 2020-09-01
 * only set `WP_ENVIRONMENT_TYPE` when changed